requirements.txt
# --- Core ---
streamlit>=1.36,<2.0
openai>=1.40,<2.0

# --- Tools / Data ---
tavily-python>=0.3
yfinance>=0.2.52
pandas>=2.2,<3.0
python-dotenv>=1.0
requests>=2.31
httpx>=0.27
pytz>=2024.1
pymongo>=4.6